fix: Fix the logic of temporary path creation
This change are made to ensure the parent directory of temporary path is exist before creating temporary files or directories. Previously, the function creates the temporary directory and then the temporary file with the same name at the same process, this also cause the creation of temporary file with no extension to failure.